+++ contrib/envconfig/CHANGELOG.md
@@ -10,3 +10,7 @@ or Security.
 # Changelog
 
 ## [Unreleased]
+
+### Breaking Changes
+
+- Raised the minimum supported Go version from 1.25.4 to 1.26.0.